Paper Prototype Concept
Basically the site is composed of three blocks. First block called 1st
floor is a header that contains a nice video, and some buttons calling
to action. 2nd floor is to register your self, login etc. 3d floor is
for replicating, 4th for becoming a developer 5th for starting a
business.